Removed cross-product from IRS calculations.
git-svn-id: https://subversion.assembla.com/svn/mdss@5358 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 6755ad4d95a7831da068ecbe6cb68b407a0f781b [formerly 3e37aa8eaf13ba68722ae11d0745025cc8ee9d7d]
Former-commit-id: 3945ac55550e75a31bd0cd04be60bd7f31948021

+++ trunk/src/server/dss/vector/solutions/irs/IRSUnionIF.java
@@ -6,7 +6,7 @@ public interface IRSUnionIF extends Reloadable
 {
   static final String RESOURCE_TARGET_VIEW = "resourceTargetView";
   
-  public static final String INTEGER = "integer";
+  public static final String FLOAT = "float";
   public static final String TEXT = "text";
   public static final String DATE = "date";
